Fixups via eclipse.
imported from: dustin@spy.net--projects-2005/photo--main--4.0--patch-17

+++ src/java/net/spy/photo/util/PhotoBackup.java
@@ -168,7 +168,7 @@ public class PhotoBackup extends Object {
 
 		public String getStats() {
 			double avgProcessTime=((double)totalTime/(double)done)/1000.0;
-			double estimate=avgProcessTime*(double)left;
+			double estimate=avgProcessTime*left;
 
 			java.text.NumberFormat nf=java.text.NumberFormat.getInstance();
 			nf.setMaximumFractionDigits(2);
